reordenando las relaciones

+++ src/modules/beneficiary/services/beneficiary.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { Injectable, NotFoundException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Beneficiary } from "../entities/beneficiary.entity";
 import { Repository } from "typeorm";
@@ -22,26 +22,37 @@ export class BeneFiciaryService{
             return newBeneficiary;
         }
         catch(error){
-            throw new Error(`Error: ${error.message}`);
+            if (error.code === '23505'){
+                throw new NotFoundException('Ya existe un beneficiario con ese nombre');
+            }
         }
     }
 
     async finAll(): Promise<Beneficiary[]>{
         try{
-           return await this.beneficiaryRepository.find({
-                relations: ['account']
-              });
+            const beneficiaries = await this.beneficiaryRepository.find({relations: ['account']});
+            if (!beneficiaries) throw new NotFoundException('No se encontraron beneficiarios');
+            return beneficiaries;
+
         }
         catch(error){
-            throw new Error(`Error: ${error.message}`);
+            if (error instanceof NotFoundException){
+                throw new NotFoundException(error.message);
+               
+            }
         }
     }
     async findOne(id: number): Promise<Beneficiary>{
         try{
-            return await this.beneficiaryRepository.findOne({where:{id}});
+            const beneficiary = await this.beneficiaryRepository.findOne({where:{id}, relations: ['account']});
+            if (!beneficiary) throw new NotFoundException('No se encuetra el beneficiario');
+            return beneficiary;
         }
         catch(error){
-            throw new Error(`Error: ${error.message}`);
+            if (error instanceof NotFoundException){
+                throw new NotFoundException(error.message);
+               
+            }
         }
     }
 
@@ -58,11 +69,14 @@ export class BeneFiciaryService{
     async delete(id: number): Promise<Beneficiary>{
         try{
             const beneficiary = await this.beneficiaryRepository.findOne({where:{id}});
+            if(!beneficiary) throw new NotFoundException(`no existe el beneficiario con id ${beneficiary.name}`)
             await this.beneficiaryRepository.delete(id);
             return beneficiary;
         }
         catch(error){
-            throw new Error(`Error: ${error.message}`);
+            if(error instanceof NotFoundException){
+                throw new NotFoundException(error.message)
+            }
         }
     }
 }
